chore: Update internationalization module configuration
This commit updates the configuration of the internationalization module in the `libs/common/src/internationalization/internationalization.module.ts` file. It adds the `ConfigModule` as an import and sets the `path` option to `path.join(__dirname,"/src/i18n/")`. It also updates the resolvers to include the `QueryResolver`, `AcceptLanguageResolver`, and `HeaderResolver` with the appropriate options.

```