chore: Update .env file configurations for apps and libs
This commit includes changes to the .gitignore file to include all .env files in the apps and libs directories. It also updates the envFilePath configuration in the authentication app module to include the .env.production.example file. These changes ensure that the correct environment variables are loaded for the respective apps and libs.

* IMPORTANT : this is only example repo none of .env file will store in inside of project