chore: Update MongoDB and RabbitMQ configurations
This commit updates the MongoDB and RabbitMQ configurations in the `.env.development` and `.env.production.example` files. It sets the `MONGODB_URI` to `"mongodb://root:example@localhost:27017/?authMechanism=DEFAULT"` and the `MONGO_DB_NAME` to `"nest-ecommerce"`. It also updates the `RABBIT_MQ_URI` to `"amqp://rabbitmq:rabbitmq@127.0.0.1:5672/vhost"`. These changes ensure that the correct database and message queue connections are used in the API gateway.